% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvrddetailserver.R
\name{jhhrvrddetailServer}
\alias{jhhrvrddetailServer}
\title{Title 后台处理总函数}
\usage{
jhhrvrddetailServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
Title 后台处理总函数
}
\examples{
HrvServer()
}
