% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvserver.R
\name{reuploadserver}
\alias{reuploadserver}
\title{Title 重新上传数据}
\usage{
reuploadserver(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
Title 重新上传数据
}
\examples{
reuploadserver()
}
