% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvserver.R
\name{revoucherserver}
\alias{revoucherserver}
\title{Title 重新生成凭证}
\usage{
revoucherserver(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
Title 重新生成凭证
}
\examples{
revoucherserver()
}
