% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvacctserver.R
\name{viewacctserver}
\alias{viewacctserver}
\title{Title 预览数据}
\usage{
viewacctserver(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
Title 预览数据
}
\examples{
viewacctserver()  
}
